Vinisha, Rakesh | MOBN-1382 | Refactor test folder structure
